+++ CHANGELOG.md
@@ -6,8 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+ - ZoomRoom Call Record console command
  - ZoomRoom Call Lock console command
  - Zoom Call Lock Converter/Response to Call Configuration Converter/Response
+ - Zoom Updated Call Record Info Event Converter/Response
 
 ## [13.8.1] - 2019-10-08
 ### Changed